fix: alterando dados

+++ database-structures/contractor_offers_service.sql
@@ -1,6 +1,6 @@
 CREATE TABLE PRESTADOR_OFERTA_SERVICO (
     cnpj_prestador VARCHAR(14) NOT NULL,
-    id_servico INT NOT NULL,
+    id_servico BIGINT NOT NULL,
     quantidade INT NOT NULL,
     data_oferta_servico DATE NOT NULL,
     PRIMARY KEY (cnpj_prestador, id_servico),
